added menu feature

+++ src/main/java/org/launchcode/controllers/CheeseController.java
@@ -8,10 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
@@ -50,6 +47,7 @@ public class CheeseController {
     public String processAddCheeseForm(@ModelAttribute @Valid Cheese newCheese, Errors errors, Model model, @RequestParam int categoryId) {
         if (errors.hasErrors()) {
             model.addAttribute("title", "Add Cheese");
+            model.addAttribute("categories", categoryDao.findAll());
             return "cheese/add";
         }
 
@@ -68,12 +66,20 @@ public class CheeseController {
 
     @RequestMapping(value = "remove", method = RequestMethod.POST)
     public String processRemoveCheeseForm(@RequestParam int[] cheeseIds) {
-
         for (int cheeseId : cheeseIds) {
-            cheeseDao.delete(cheeseId);
+            if (cheeseId > 0) {
+                cheeseDao.delete(cheeseId);
+            }
         }
 
         return "redirect:";
     }
 
+    @RequestMapping(value = "category/{id}", method = RequestMethod.GET)
+    public String category(@PathVariable int id, Model model) {
+        Category c = categoryDao.findOne(id);
+        model.addAttribute("cheeses", c.getCheeses());
+        model.addAttribute("title", "Cheeses in Category: " + c.getName());
+        return "cheese/index";
+    }
 }
